Update MIME database at image build time (if requested by some puavo-pkg configuration at build time), or after boot as a systemd service (if requested by some puavo-pkg configuration at boot time).
Previously this would work only for the "smartboard" puavo-pkg,
but it would have to be installed as part of the image to not slow
down boot considerably.  Now this should work for other packages that
need it, as well as "smartboard", which can now also be installed as a
separate package for laptops and fatclients (this also fixes the error
with read-only upstream directories that would otherwise happen with
"smartboard" on fatclients).