Revert "Remove Teams, the standalone app for Linux has been retired by"
This reverts commit 2e081596740e53edf986b7bfc3c88fab2d40e92c.

Do not do this quite yet.  Microsoft is not distributing this
anymore, but they are still actively suggesting to use this.
Currently the PWA can not be installed in Google Chrome (or
rather, it does not show a button to do so), even though other
PWAs work.  Microsoft is confused and we will do something once
it is clear what should be done.