Add puavo-pkg for Kantti, an educational game about substance abuse and mental health.
This can not be used yet.  The binary is actually an updater
that downloads a huge file (730MB at the moment) into user's
home directory.  On computers which are used by many people this
leads to waste of space which is probably a problem when files
are as big as this.  We need to figure out if should be solve this
problem and if yes, how could we do that.